+++ modules/Custom/RemoveAccess.php
@@ -31,8 +31,14 @@ if ( $_REQUEST['modfunc'] === 'save'
 
 		if ( $accessfunc === 'grant' )
 		{
+			// @since 11.0 SQL Fix replacement in case the "username" contains the prefix
 			$update_sql = "UPDATE students
-				SET USERNAME=REPLACE(USERNAME,'" . $username_prefix_add . "','')
+				SET USERNAME=CONCAT(
+					REPLACE(
+						SUBSTRING(USERNAME, 1, LENGTH('" . $username_prefix_add . "')),
+						'" . $username_prefix_add . "',
+						''),
+					SUBSTR(USERNAME, LENGTH('" . $username_prefix_add . "')+1, 100))
 				WHERE STUDENT_ID IN (" . $st_list . ")";
 		}